feat: classname composition component
