+++ components/ui/ui-builder/internal/classname-control/index.tsx
@@ -1,0 +1,6 @@
+"use client";
+import { BreakpointClassNameControl } from "@/components/ui/ui-builder/internal/classname-control/breakpoint-classname-control";
+import { BreakpointClassNameField } from "@/components/ui/ui-builder/internal/classname-control/breakpoint-classname-control-field";
+
+export { BreakpointClassNameField as ClassNameField };
+export default BreakpointClassNameControl;
