+++ components/ui/ui-builder/internal/classname-control/utils.tsx
@@ -1,0 +1,7 @@
+// Type guards for Tailwind class arrays
+export function isTailwindClass<T extends readonly string[]>(
+    arr: T,
+    token: string
+  ): token is T[number] {
+    return arr.includes(token as any);
+  }
